Move DownloadReportDialog and TriggerAlertDialog to /pages/reports
